geotiff tests: consolidate rotated-CRS cluster (#2396)
Folds three rotated / dropped-CRS test files into a single parametrized
file at ``read/test_crs.py`` under the layout established by epic #2390
PR 1:

* ``test_allow_rotated_geotiff_2115.py``
* ``test_allow_rotated_crs_drop_2126.py``
* ``test_allow_rotated_no_crs_2122.py``

Parametrizes scenarios as ``rotated_no_crs``, ``rotated_with_crs``,
``axis_aligned_with_crs`` with ``eager`` / ``dask`` backend axes. The
GPU eager + dask+CuPy cases stay gated by ``@requires_gpu``.

The two cross-file ``_write_rotated_tiff`` consumers
(``test_georef_status_2136.py``, ``test_lazy_finalization_parity_2162.py``)
now import from ``read/test_crs.py``. The release-gate and reference
docs that cite the old filenames are updated to point at the new file.

The HTTP rotated test (``test_http_dask_allow_rotated_2130.py``) is
left in place; it belongs to PR 9 (integration cluster).

``CLUSTER_AUDIT_PR3.md`` maps every old ``file::test`` to its new
``file::test_id`` and is removed in a follow-up commit before approval.

Part of epic #2390.
