+++ xrspatial/geotiff/tests/read/__init__.py
@@ -1,0 +1,1 @@
+"""Read-path tests for the GeoTIFF module."""
